fix(download): wait for response end
The `DropboxService` was not handling calls to `https.get` appropriately
when downloading files.  The code was resolving the promise after
receiving a `close` event from the request.  This could result in the
promise resolving before the file had finished downloading. In addition,
the code was not verifying the status code of the response was a
successful code.

Changes:
* Resolve the promise when the request emits an `end` event and not when
  the response emits a `close` event.
* If the response status code is not a successful code then reject the
  promise.